moved deliveryAction back into DeliveryTrait
